chore: move related files to dockest package
