+++ package.json
@@ -6,25 +6,12 @@
     "packages/dockest",
     "packages/examples/*"
   ],
-  "description": "Dockest is an integration testing tool aimed at alleviating the process of evaluating unit tests whilst running multi-container Docker applications.",
-  "keywords": [
-    "docker",
-    "docker-compose",
-    "jest",
-    "nodejs",
-    "node",
-    "integration testing"
-  ],
   "homepage": "https://github.com/erikengervall/dockest#readme",
   "bugs": {
     "url": "https://github.com/erikengervall/dockest/issues"
   },
   "license": "MIT",
   "author": "Erik Engervall <erik@engervall.com> (http://engervall.com)",
-  "files": [
-    "dist/**/*"
-  ],
-  "main": "dist/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/erikengervall/dockest.git"
