+++ packages/dockest/fixtures/docker-compose-single-redis-duplicate.yml
